minimal hosting model
