--- examples/MicroserviceExample/WebApi/Startup.cs
@@ -1,49 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-using OpenTelemetry.Trace;
-using Utils.Messaging;
-
-namespace WebApi;
-
-public class Startup
-{
-    public Startup(IConfiguration configuration)
-    {
-        this.Configuration = configuration;
-    }
-
-    public IConfiguration Configuration { get; }
-
-    public void ConfigureServices(IServiceCollection services)
-    {
-        services.AddControllers();
-
-        services.AddSingleton<MessageSender>();
-
-        services.AddOpenTelemetry()
-            .WithTracing(builder => builder
-                .AddAspNetCoreInstrumentation()
-                .AddSource(nameof(MessageSender))
-                .AddZipkinExporter(b =>
-                {
-                    var zipkinHostName = Environment.GetEnvironmentVariable("ZIPKIN_HOSTNAME") ?? "localhost";
-                    b.Endpoint = new Uri($"http://{zipkinHostName}:9411/api/v2/spans");
-                }));
-    }
-
-    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-    {
-        if (env.IsDevelopment())
-        {
-            app.UseDeveloperExceptionPage();
-        }
-
-        app.UseRouting();
-
-        app.UseEndpoints(endpoints =>
-        {
-            endpoints.MapControllers();
-        });
-    }
-}
